requirements.txt
PyQt5==5.15.10
PyYAML==6.0.1
ruamel.yaml==0.18.6
